+++ packages/twenty-front/src/modules/workflow/workflow-trigger/utils/cron-to-human/descriptors/getMonthsDescription.ts
@@ -2,12 +2,10 @@ import { t } from '@lingui/core/macro';
 import { format, type Locale } from 'date-fns';
 import { isDefined } from 'twenty-shared/utils';
 
+import { isListValue } from '~/utils/validation/isListValue';
+import { isNumericRange } from '~/utils/validation/isNumericRange';
+import { isStepValue } from '~/utils/validation/isStepValue';
 import { type CronDescriptionOptions } from '../types/cronDescriptionOptions';
-import {
-  isListValue,
-  isNumericRange,
-  isStepValue,
-} from '../utils/cronStringUtilities';
 
 const getMonthName = (
   monthNum: number,
