+++ packages/twenty-front/src/utils/format/__tests__/getOrdinalNumber.test.ts
@@ -1,0 +1,14 @@
+import { getOrdinalNumber } from '../getOrdinalNumber';
+
+describe('getOrdinalNumber', () => {
+  it('should return correct ordinal numbers', () => {
+    expect(getOrdinalNumber(1)).toBe('1st');
+    expect(getOrdinalNumber(2)).toBe('2nd');
+    expect(getOrdinalNumber(3)).toBe('3rd');
+    expect(getOrdinalNumber(4)).toBe('4th');
+    expect(getOrdinalNumber(11)).toBe('11th');
+    expect(getOrdinalNumber(21)).toBe('21st');
+    expect(getOrdinalNumber(22)).toBe('22nd');
+    expect(getOrdinalNumber(23)).toBe('23rd');
+  });
+});
