+++ packages/twenty-front/src/utils/validation/isListValue.ts
@@ -1,0 +1,5 @@
+import { isDefined } from 'twenty-shared/utils';
+
+export const isListValue = (value: string): boolean => {
+  return isDefined(value) && value.includes(',');
+};
