+++ packages/twenty-front/src/utils/validation/isNumericRange.ts
@@ -1,0 +1,5 @@
+import { isDefined } from 'twenty-shared/utils';
+
+export const isNumericRange = (value: string): boolean => {
+  return isDefined(value) && /^\d+(-\d+)?$/.test(value);
+};
