+++ packages/twenty-front/src/utils/validation/isStepValue.ts
@@ -1,0 +1,5 @@
+import { isDefined } from 'twenty-shared/utils';
+
+export const isStepValue = (value: string): boolean => {
+  return isDefined(value) && value.includes('/');
+};
